Refactor topic creation lambda for concurrent use
The subscriber lists are now sets so updates like insertion and deletion of
subscribers within these sets can happen within one request to the
database and supposedly atomically. (With a list, deletion of an item
seems to happen by index which is not suited for lambdas concurrently
executing and modifying the list).
An empty set is not a valid attribute value, so to remove further
implementation messiness of sometimes having an attribute and sometimes
not having it - which would bring more concurrency issues, I chose to
add a dummy object in the set which would always make sure it's not
empty.
Also the check if item already exists is now in the put_item request
which again helps with concurrency and also makes it useless for the
lambda to have query permissions.