Uniform the lambda parameters expected
The main lambda will receive one parameter 'parameters' - a JSON object.
It will extract its properties and pass it to the helper lambdas.
The only required of those parameters is the 'message' property.
It should be a JSON object and can optionally contain 'subject' and 'body' properties.
All the other parameters are optional and can be used for custom configuration of the different AWS services.